refactor initial page object spec
