+++ spec/integration/pages/editable.rb
@@ -1,0 +1,2 @@
+class Editable < SitePrism::Section
+end
