+++ spec/integration/pages/editor.rb
@@ -1,0 +1,4 @@
+require_relative 'editable'
+class Editor < SitePrism::Page
+  section :editable, Editable, '#editable1'
+end
