+++ spec/spec_helper.rb
@@ -38,6 +38,7 @@ RSpec.configure do |config|
 
   config.before(:each, js: true) do
     DatabaseCleaner.strategy = :truncation
+    require 'integration/support/active_controller'
   end
 
   config.before(:each) do
